chore: add jitter only after authorization
