+++ backend/src/email/routes/email-callback.routes.ts
@@ -1,5 +1,6 @@
 import { Router } from 'express'
 import { EmailCallbackMiddleware } from '@email/middlewares'
+import { jitter } from '@core/utils/request'
 const router = Router()
 /**
  * @swagger
@@ -19,6 +20,7 @@ router.post(
   '/',
   EmailCallbackMiddleware.printConfirmSubscription,
   EmailCallbackMiddleware.isAuthenticated,
+  jitter,
   EmailCallbackMiddleware.parseEvent
 )
